+++ src/renderer/features/flexible-multisig-create/index.ts
@@ -1,2 +1,2 @@
 export { FlexibleMultisigWallet } from './components/FlexibleMultisigWallet';
-export { flexibleMultisigModel } from './model/flexible-multisig';
+export { flexibleMultisigModel } from './model/flexible-multisig-create';
