+++ src/renderer/features/flexible-multisig-create/model/__tests__/confirm-model.test.ts
