+++ src/renderer/features/flexible-multisig-create/model/__tests__/mock.ts
