+++ src/renderer/features/flexible-multisig-create/model/__tests__/signatory-model.test.ts
