+++ src/renderer/features/flexible-multisig-create/model/confirm-model.ts
