+++ src/renderer/features/flexible-multisig-create/model/wallet-provider-model.ts
